Move things around, add tests and whitelists
