+++ src/consumer.tsx
@@ -1,5 +1,5 @@
-import I18nContext from './context'
+import { I18nContext } from './context'
 
 const I18nConsumer = I18nContext.Consumer
 
-export default I18nConsumer
+export { I18nConsumer }
