+++ src/index.ts
@@ -1,20 +1,6 @@
-export { default as I18nContext, I18nContextProps } from './context'
-export { default as I18nProvider } from './provider'
-export { default as I18nConsumer } from './consumer'
-export { default as withTranslate } from './hoc'
-export {
-  default as I18nStore,
-  I18nStoreState,
-  I18nStoreCallback,
-} from './store'
-export {
-  TranslateData,
-  TranslateDataWithJSX,
-  TranslateFunc,
-  TranslateJsxFunc,
-  TranslatePluralFunc,
-  TranslatePluralJsxFunc,
-  TranslationOptions,
-  TranslationSet,
-  Translations,
-} from './types'
+export * from './context'
+export * from './provider'
+export * from './consumer'
+export * from './hoc'
+export * from './store'
+export * from './types'
