+++ src/parser/index.ts
@@ -1,5 +1,5 @@
 export { SyntaxKind, Token, DocNode } from './types'
-export { default as ElementNode } from './ElementNode'
-export { default as TextNode } from './TextNode'
-export { default as voidElements } from './void-elements'
-export { default as parse } from './parse'
+export { ElementNode } from './ElementNode'
+export { TextNode } from './TextNode'
+export { voidElements } from './void-elements'
+export { parse } from './parse'
