+++ src/translate/index.tsx
@@ -1,0 +1,3 @@
+export { getTranslation } from './translate'
+export { replaceJsx, replaceString } from './replace-content'
+export { normalizeContent } from './normalize-content'
